--- client/components/signIn.jsx
@@ -1,12 +1,0 @@
-import React, { useState } from 'react';
-
-function SignIn() {
-
-  return (
-    <div>
-      Sign in Component
-    </div>
-  )
-}
-
-export default SignIn;
