+++ src/index.js
@@ -2,4 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from '../client/components/App.jsx';
 
+// import Router from '../client/components/AppRouter.jsx/index.js';
+
+
 ReactDOM.render(<App />, document.getElementById('root'));
